Revert screen names for Google Analytics
After the publication of the new version of the app, the screen names
in google analytics didn’t match with the old ones, so..the names were
restored and moved to the previous package: com.permutas.sep.ui. This
way the reports will need todo minimal adjustments in order to analyze
the app usage statistics.